+++ scripts/macros/2024/classFeatures/ranger/feyWanderer/otherworldlyGlamour.js
@@ -1,0 +1,5 @@
+export let otherworldlyGlamour = {
+    name: 'Otherworldly Glamour',
+    version: '1.3.99',
+    rules: 'modern'
+};