+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.3.99 Release:
 ## New Modern Content:
 - Dreadful Strikes
+- Otherworldly Glamour
 - Beguiling Twist
 - Fey Reinforcements
 - Misty Wanderer